+++ Lib/fontbakery/specifications/googlefonts.py
@@ -100,7 +100,7 @@ expected_check_ids = \
       , 'com.google.fonts/check/vendor_id'
       , 'com.google.fonts/check/019' # Substitute copyright, registered and trademark symbols in name table entries.
       , 'com.google.fonts/check/usweightclass'
-      , 'com.google.fonts/check/028' # Check font has a license.
+      , 'com.google.fonts/check/has_license'
       , 'com.google.fonts/check/029' # Check copyright namerecords match license file.
       , 'com.google.fonts/check/030' # License URL matches License text on name table?
       , 'com.google.fonts/check/name/no_copyright_on_description'
@@ -822,9 +822,9 @@ def license(license_path):
 
 
 @check(
-  id = 'com.google.fonts/check/028'
+  id = 'com.google.fonts/check/has_license'
 )
-def com_google_fonts_check_028(licenses):
+def com_google_fonts_check_has_license(licenses):
   """Check font has a license."""
   if len(licenses) > 1:
     yield FAIL, Message("multiple",